backend/requirements.txt
astroid==2.2.5
Click==7.0
ecdsa==0.13.2
Flask
Flask-SQLAlchemy
future==0.17.1
isort==4.3.18
itsdangerous==1.1.0
lazy-object-proxy==1.4.0
MarkupSafe==1.1.1
mccabe==0.6.1
pycryptodome==3.3.1
pylint==2.3.1
python-jose[cryptography]
six==1.12.0
typed-ast==1.3.5
wrapt==1.11.1
Flask-Cors